renaming queryctx to driverctx
