+++ go.mod
@@ -28,7 +28,6 @@ require (
 )
 
 require (
-	github.com/google/uuid v1.3.0
 	github.com/pkg/errors v0.9.1
 	github.com/rs/zerolog v1.28.0
 	golang.org/x/sys v0.0.0-20220811171246-fbc7d0a398ab // indirect
